move shouldFailWithMessage test helper to appropriate location
