+++ src/helpers/test-utils/functions/index.ts
@@ -1,2 +1,3 @@
 export * from './conversions';
 export * from './expect';
+export * from './shouldFailWithMessage';
